requirements.txt
Django==3.2.5
djangorestframework==3.12.4
drf-yasg==1.20.0
django-filter
psycopg2-binary==2.8.6
gunicorn==20.1.0
pytest==6.2.4
pytest-django==4.3.0
